enable DTmap calls for ATVApaynt
# Conflicts:
#	paynt/quotient/mdp.py
#	paynt/synthesizer/decision_tree.py

# Conflicts:
#	paynt/synthesizer/decision_tree.py
