+++ paynt/quotient/mdp_family.py
@@ -3,12 +3,13 @@ import payntbind
 import paynt.family.family
 import paynt.quotient.quotient
 import paynt.models.models
+import paynt.quotient.utils.variable
+import paynt.quotient.utils.decision_tree
 
 import json
 
 import logging
 logger = logging.getLogger(__name__)
-
 
 class MdpFamilyQuotient(paynt.quotient.quotient.Quotient):
 
@@ -55,6 +56,24 @@ class MdpFamilyQuotient(paynt.quotient.quotient.Quotient):
             self.quotient_mdp, self.num_actions, self.choice_to_action)
         self.state_to_actions = MdpFamilyQuotient.map_state_to_available_actions(self.state_action_choices)
 
+        # Copy from MdpQuotient class, also variable class
+        assert quotient_mdp.has_state_valuations(), "model has no state valuations"
+        sv = quotient_mdp.state_valuations
+        valuation = json.loads(str(sv.get_json(0)))
+        variable_name = [var_name for var_name in valuation]
+        state_valuations = []
+        for state in range(quotient_mdp.nr_states):
+            valuation = json.loads(str(sv.get_json(state)))
+            valuation = [valuation[var_name] for var_name in variable_name]
+            state_valuations.append(valuation)
+        variables = [paynt.quotient.utils.variable.Variable(var, var_name, state_valuations) for var, var_name in enumerate(variable_name)]
+        variable_mask = [len(v.domain) > 1 for v in variables]
+        variables = [v for index, v in enumerate(variables) if variable_mask[index]]
+        for state, valuation in enumerate(state_valuations):
+            state_valuations[state] = [value for index, value in enumerate(valuation) if variable_mask[index]]
+        self.variables = variables
+        self.state_valuations = state_valuations
+
     def empty_policy(self):
         return self.empty_scheduler()
 
@@ -66,6 +85,76 @@ class MdpFamilyQuotient(paynt.quotient.quotient.Quotient):
             if choice is not None:
                 policy[state] = self.choice_to_action[choice]
         return policy
+
+    def state_valuation_to_state(self, valuation):
+        valuation = [valuation[v.name] for v in self.variables]
+        for state,state_valuation in enumerate(self.state_valuations):
+            if valuation == state_valuation:
+                return state
+        else:
+            assert False, "state valuation not found"
+
+    def policy_to_policy_vector(self, policy):
+        # assert self.quotient_mdp.nr_states == len(policy), "policy length does not match the number of states"
+        choices = self.state_to_choice_to_choices(policy)
+        return choices
+
+
+    def reset_tree(self, depth, enable_harmonization=True):
+        '''
+        Rebuild the decision tree template, the design space and the coloring.
+        '''
+        logger.debug(f"building tree of depth {depth}")
+        self.decision_tree = paynt.quotient.utils.decision_tree.DecisionTree(self,self.variables,self.state_valuations)
+        self.decision_tree.set_depth(depth)
+
+        variables = self.decision_tree.variables
+        variable_name = [v.name for v in variables]
+        variable_domain = [v.domain for v in variables]
+        tree_list = self.decision_tree.to_list()
+        self.coloring = payntbind.synthesis.ColoringSmt(
+            self.quotient_mdp.nondeterministic_choice_indices, self.choice_to_action, self.quotient_mdp.state_valuations,
+            variable_name, variable_domain, tree_list, enable_harmonization
+        )
+        self.coloring.enableStateExploration(self.quotient_mdp)
+
+        # reconstruct the family
+        hole_info = self.coloring.getFamilyInfo()
+        self.family = paynt.family.family.Family()
+        self.is_action_hole = [False for hole in hole_info]
+        self.is_decision_hole = [False for hole in hole_info]
+        self.is_variable_hole = [False for hole in hole_info]
+        domain_max = max([len(domain) for domain in variable_domain])
+        bound_domain = list(range(domain_max))
+        node_hole_info = [[] for node in self.decision_tree.collect_nodes()]
+        for hole,info in enumerate(hole_info):
+            node,hole_name,hole_type = info
+            node_hole_info[node].append( (hole,hole_name,hole_type) )
+            if hole_type == "__action__":
+                self.is_action_hole[hole] = True
+                option_labels = self.action_labels
+            elif hole_type == "__decision__":
+                self.is_decision_hole[hole] = True
+                option_labels = variable_name
+            else:
+                self.is_variable_hole[hole] = True
+                variable = variable_name.index(hole_type)
+                option_labels = variables[variable].hole_domain
+            self.family.add_hole(hole_name, option_labels)
+        self.splitter_count = [0] * self.family.num_holes
+        self.decision_tree.root.associate_holes(node_hole_info)
+
+    def build_unsat_result(self):
+        spec_result = paynt.verification.property_result.MdpSpecificationResult()
+        spec_result.constraints_result = paynt.verification.property_result.ConstraintsResult([])
+        spec_result.optimality_result = paynt.verification.property_result.MdpOptimalityResult(None)
+        spec_result.evaluate(None)
+        spec_result.can_improve = False
+        return spec_result
+
+    def are_choices_consistent(self, choices, family):
+        ''' Separate method for profiling purposes. '''
+        return self.coloring.areChoicesConsistent(choices, family.family)
 
     def policy_to_state_valuation_actions(self, policy):
         '''
@@ -199,4 +288,5 @@ class MdpFamilyQuotient(paynt.quotient.quotient.Quotient):
         assert family.size == 1, "expecting family of size 1"
         choices = self.coloring.selectCompatibleChoices(family.family)
         model,state_map,choice_map = self.restrict_quotient(choices)
+        # model = MdpFamilyQuotient.mdp_to_dtmc(model)
         return paynt.models.models.SubMdp(model,state_map,choice_map)
