+++ paynt/quotient/utils/__init__.py
@@ -1,0 +1,10 @@
+__version__ = "unknown"
+
+try:
+    from ._version import __version__
+except ImportError:
+    # We're running in a tree that doesn't have a _version.py, so we don't know what our version is.
+    pass
+
+def version():
+    return __version__
