+++ paynt/quotient/utils/decision_tree.py
@@ -1,0 +1,199 @@
+import logging
+logger = logging.getLogger(__name__)
+
+class DecisionTree:
+
+    def __init__(self, quotient, variables, state_valuations):
+        self.quotient = quotient
+        self.state_valuations = state_valuations
+        self.variables = variables
+        self.reset()
+
+    def reset(self):
+        self.root = DecisionTreeNode(None)
+
+    def set_depth(self, depth:int):
+        self.reset()
+        for level in range(depth):
+            for node in self.collect_terminals():
+                node.add_children()
+        self.root.assign_identifiers()
+
+    def get_depth(self):
+        return self.root.get_depth()
+
+    def collect_nodes(self, node_condition=None):
+        if node_condition is None:
+            node_condition = lambda node : True
+        node_queue = [self.root]
+        output_nodes = []
+        while node_queue:
+            node = node_queue.pop(0)
+            if node_condition(node):
+                output_nodes.append(node)
+            node_queue += node.child_nodes
+        return output_nodes
+
+    def collect_terminals(self):
+        return self.collect_nodes(lambda node : node.is_terminal)
+
+    def collect_nonterminals(self):
+        return self.collect_nodes(lambda node : not node.is_terminal)
+
+    def to_list(self):
+        num_nodes = len(self.collect_nodes())
+        node_info = [ None for node in range(num_nodes) ]
+        for node in self.collect_nodes():
+            parent = num_nodes if node.parent is None else node.parent.identifier
+            child_true = num_nodes if node.is_terminal else node.child_true.identifier
+            child_false = num_nodes if node.is_terminal else node.child_false.identifier
+            node_info[node.identifier] = (parent,child_true,child_false)
+        return node_info
+
+    def simplify(self, target_state_mask):
+        state_valuations = [self.state_valuations[state] for state in ~target_state_mask]
+        self.root.simplify(self.variables, state_valuations)
+
+    def to_string(self):
+        return self.root.to_string(self.variables,self.quotient.action_labels)
+
+    def to_graphviz(self):
+        import graphviz
+        logging.getLogger("graphviz").setLevel(logging.WARNING)
+        logging.getLogger("graphviz.sources").setLevel(logging.ERROR)
+        graphviz_tree = graphviz.Digraph(comment="decision tree")
+        self.root.to_graphviz(graphviz_tree,self.variables,self.quotient.action_labels)
+        return graphviz_tree
+
+
+class DecisionTreeNode:
+
+    def __init__(self, parent):
+        self.parent = parent
+        self.child_true = None
+        self.child_false = None
+        self.identifier = None
+        self.holes = None
+
+        self.action = None
+        self.variable = None
+        self.variable_bound = None
+
+    @property
+    def is_terminal(self):
+        return self.child_false is None and self.child_true is None
+
+    @property
+    def child_nodes(self):
+        return [] if self.is_terminal else [self.child_true,self.child_false]
+
+    @property
+    def is_true_child(self):
+        return self is self.parent.child_true
+
+    def add_children(self):
+        assert self.is_terminal
+        self.child_true = DecisionTreeNode(self)
+        self.child_false = DecisionTreeNode(self)
+
+    def get_depth(self):
+        if self.is_terminal:
+            return 0
+        return 1 + max([child.get_depth() for child in self.child_nodes])
+
+    def assign_identifiers(self, identifier=0):
+        self.identifier = identifier
+        if self.is_terminal:
+            return self.identifier
+        identifier = self.child_true.assign_identifiers(identifier+1)
+        identifier = self.child_false.assign_identifiers(identifier+1)
+        return identifier
+
+    def associate_holes(self, node_hole_info):
+        self.holes = [hole for hole,_,_ in node_hole_info[self.identifier]]
+        if self.is_terminal:
+            return
+        self.child_true.associate_holes(node_hole_info)
+        self.child_false.associate_holes(node_hole_info)
+
+    def associate_assignment(self, assignment):
+        hole_assignment = [assignment.hole_options(hole)[0] for hole in self.holes]
+        if self.is_terminal:
+            self.action = hole_assignment[0]
+            return
+
+        self.variable = hole_assignment[0]
+        self.variable_bound = hole_assignment[self.variable+1]
+
+        self.child_true.associate_assignment(assignment)
+        self.child_false.associate_assignment(assignment)
+
+    def apply_hint(self, subfamily, tree_hint):
+        if self.is_terminal or tree_hint.is_terminal:
+            return
+
+        variable_hint = tree_hint.variable
+        subfamily.hole_set_options(self.holes[0],[variable_hint])
+        subfamily.hole_set_options(self.holes[variable_hint+1],[tree_hint.variable_bound])
+        self.child_true.apply_hint(subfamily,tree_hint.child_true)
+        self.child_false.apply_hint(subfamily,tree_hint.child_false)
+
+    def simplify(self, variables, state_valuations):
+        if self.is_terminal:
+            return
+
+        bound = variables[self.variable].domain[self.variable_bound]
+        state_valuations_true =  [valuation for valuation in state_valuations if valuation[self.variable] <= bound]
+        state_valuations_false = [valuation for valuation in state_valuations if valuation[self.variable]  > bound]
+        child_skip = None
+        if len(state_valuations_true) == 0:
+            child_skip = self.child_false
+        elif len(state_valuations_false) == 0:
+            child_skip = self.child_true
+        if child_skip is not None:
+            self.variable = child_skip.variable
+            self.variable_bound = child_skip.variable_bound
+            self.action = child_skip.action
+            self.child_true = child_skip.child_true
+            self.child_false = child_skip.child_false
+            self.simplify(variables,state_valuations)
+            return
+
+        self.child_true.simplify(variables, state_valuations_true)
+        self.child_false.simplify(variables, state_valuations_false)
+        if not self.is_terminal and self.child_true.is_terminal and self.child_false.is_terminal and self.child_true.action == self.child_false.action:
+            self.variable = self.variable_bound = None
+            self.action = self.child_true.action
+            self.child_true = self.child_false = None
+
+    def to_string(self, variables, action_labels, indent_level=0, indent_size=2):
+        indent = " "*indent_level*indent_size
+        if self.is_terminal:
+            return indent + f"{action_labels[self.action]}" + "\n"
+        var = variables[self.variable]
+        s = ""
+        s += indent + f"if {var.name}<={var.domain[self.variable_bound]}:" + "\n"
+        s += self.child_true.to_string(variables,action_labels,indent_level+1)
+        s += indent + f"else:" + "\n"
+        s += self.child_false.to_string(variables,action_labels,indent_level+1)
+        return s
+
+    @property
+    def graphviz_id(self):
+        return str(self.identifier)
+
+    def to_graphviz(self, graphviz_tree, variables, action_labels):
+        if not self.is_terminal:
+            for child in self.child_nodes:
+                child.to_graphviz(graphviz_tree,variables,action_labels)
+
+        if self.is_terminal:
+            node_label = action_labels[self.action]
+        else:
+            var = variables[self.variable]
+            node_label = f"{var.name}<={var.domain[self.variable_bound]}"
+
+        graphviz_tree.node(self.graphviz_id, label=node_label, shape="box", style="rounded", margin="0.05,0.05")
+        if not self.is_terminal:
+            graphviz_tree.edge(self.graphviz_id,self.child_true.graphviz_id,label="T")
+            graphviz_tree.edge(self.graphviz_id,self.child_false.graphviz_id,label="F")
